refactor, cleanup
